// src/app/api/model/listAllPpoReceiptsResponseDTO.ts
/**
 * CTS-BE
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateOnly } from './dateOnly';


export interface ListAllPpoReceiptsResponseDTO { 
    id?: number;
    treasuryReceiptNo?: string | null;
    ppoNo?: string | null;
    pensionerName?: string | null;
    receiptDate?: DateOnly;
}